Sorting `config` and moving the file up a directory.
