--- src/config/index.js
@@ -1,3 +1,0 @@
-import config from './config.js';
-
-export default config;
